+++ src/pages/create_user.tsx
@@ -1,7 +1,7 @@
 import { NextPage } from 'next';
 import Head from 'next/head';
 import Layout from '../components/common/Layout';
-import CreateUserInDbHandler from '../graphql/handler/user/createUserInDbHandler';
+import CreateUserInDbHandler from '../graphql/handler/user/CreateUserInDbHandler';
 
 const CreateUser: NextPage = () => {
   return (
